Add Azure Search configuration files and update upload script
