+++ infra/definitions/skillset.json
@@ -1,0 +1,80 @@
+{
+    "name": "{INDEX_NAME}-skillset",
+    "skills": [
+        {
+            "@odata.type": "#Microsoft.Skills.Text.SplitSkill",
+            "name": "Chunk documents",
+            "description": "Split skill to chunk documents",
+            "context": "/document",
+            "defaultLanguageCode": "en",
+            "textSplitMode": "pages",
+            "maximumPageLength": 2000,
+            "pageOverlapLength": 500,
+            "maximumPagesToTake": 0,            
+            "inputs": [
+                {
+                    "name": "text",
+                    "source": "/document/content",
+                    "inputs": []
+                }
+            ],
+            "outputs": [
+                {
+                    "name": "textItems",
+                    "targetName": "pages"
+                }
+            ]
+        },
+        {
+            "@odata.type": "#Microsoft.Skills.Text.AzureOpenAIEmbeddingSkill",
+            "name": "Create embeddings",
+            "context": "/document/pages/*",
+            "resourceUri": "{AZURE_OPENAI_RESOURCE_URI}",
+            "deploymentId": "{INDEXER_EMBEDDING_MODEL_ID}",
+            "dimensions": 3072,
+            "modelName": "{EMBEDDING_MODEL_NAME}",
+            "inputs": [
+                {
+                    "name": "text",
+                    "source": "/document/pages/*",
+                    "inputs": []
+                }
+            ],
+            "outputs": [
+                {
+                    "name": "embedding",
+                    "targetName": "text_vector"
+                }
+            ]
+        }
+    ],
+    "indexProjections": {
+        "selectors": [
+            {
+                "targetIndexName": "{INDEX_NAME}",
+                "parentKeyFieldName": "parent_id",
+                "sourceContext": "/document/pages/*",
+                "mappings": [
+                    {
+                        "name": "chunk",
+                        "source": "/document/pages/*",
+                        "inputs": []
+                    },
+                    {
+                        "name": "text_vector",
+                        "source": "/document/pages/*/text_vector",
+                        "inputs": []
+                    },
+                    {
+                        "name": "title",
+                        "source": "/document/metadata_storage_name",
+                        "inputs": []
+                    }
+                ]
+            }
+        ],
+        "parameters": {
+            "projectionMode": "skipIndexingParentDocuments"
+        }
+    }
+}
